Implementação detalhes dos planos

+++ app/Http/Controllers/Admin/PlanController.php
@@ -50,10 +50,18 @@ class PlanController extends Controller
 
     public function destroy($id)
     {
-        $plan = $this->repository->where('id', $id)->get()->first();
+        $plan = $this->repository
+            ->with('details')
+            ->where('id', $id)->get()->first();
 
         if (!$plan)
             return redirect()->back();
+
+        if ($plan->details->count() > 0) {
+            return redirect()
+                ->back()
+                ->with('error', 'Existem detalhes vinculados a esse plano, portanto não é possível deletar');
+        }
 
         $plan->delete();
